# PrideHub-Backend-main/swagger.yml
openapi: 3.0.0
info:
  title: PrideHub API List
  description: This is a reference api list of PrideHub Backend.
  version: 1.0.0
  contact:
    email: yashmangal240@gmail.com
servers:
  - url: http://localhost:9999/api
    description: Development server
  - url: https://lgbtq-backend.onrender.com/api
    description: Production Server
paths:
...
  /blog/create:
    post:
      tags:
        - Blog
      summary: Create a blog
      security:
        - bearerAuth: []
      requestBody:
        $ref: '#/components/requestBodies/CreateUpdateBlog'
      responses:
        '200':
          description: Blog Created Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /blog/fetch/{blogId}:
    get:
      tags:
        - Blog
      summary: Get a particular blog
      parameters:
        - in: path
          name: blogId
          description: Blog ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Blog Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /blog/update/{blogId}:
    put:
      tags:
        - Blog
      summary: Update a blog
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: blogId
          description: Blog ID
          schema:
            type: string
            format: ObjectId
      requestBody:
        $ref: '#/components/requestBodies/CreateUpdateBlog'
      responses:
        '200':
          description: Blog Updated Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /blog/delete/{blogId}:
    delete:
      tags:
        - Blog
      summary: Delete a blog
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: blogId
          description: Blog ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Blog Deleted Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /blog/upvote/{blogId}:
    put:
      tags:
        - Blog
      summary: Upvote a blog
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: blogId
          description: Blog ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Blog Upvoted Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /blog/downvote/{blogId}:
    put:
      tags:
        - Blog
      summary: Downvote a blog
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: blogId
          description: Blog ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Blog Downvoted Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /blog/myblogs:
    get:
      tags:
        - Blog
      summary: Get all blogs created by the user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: My Blogs Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
components:
  schemas:
...
    CreateUpdateBlogInput:
      type: object
      properties:
        title:
          type: string
          required: true
        content:
          type: string
          required: true
        imageURL:
          type: string
  securitySchemes:
    bearerAuth:
      name: Authorization
      type: http
      scheme: bearer
      in: header
      bearerFormat: JWT
  requestBodies:
...
    CreateUpdateBlog:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateUpdateBlogInput'
        description: Create or Update Blog